requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlmodel==0.0.22
pillow==10.4.0
python-docx==1.1.0
fpdf2==2.7.8
PyPDF2==3.0.1
python-multipart==0.0.9
pytest==8.3.2
httpx==0.27.2
pydantic-settings==2.6.1
cairosvg==2.8.2
openpyxl==3.1.2
python-pptx==0.6.21
cffi==2.0.0
pydub==0.25.1
